+++ mobile/app/(auth)/register.js
@@ -34,7 +34,6 @@ export default function Register() {
 
   // Method to POST inputted data to /register server route
   const createUser = () => {
-    console.log("create account button clicked");
     const payload = {
       name: name,
       year: year,
@@ -45,19 +44,15 @@ export default function Register() {
 
     // Get IP that Expo server is using to host app, allows to connect with the backend
     const URI = Constants.expoConfig.hostUri.split(':').shift();
-    console.log(URI);
-    const fullUrl = `http://${URI}:5000/api/auth/register`;
-    console.log("Full URL:", fullUrl);
 
     // POST to /login with payload
-    fetch(`http://${URI}:5000/api/auth/register`, {
+    fetch(`http://${URI}:${process.env.EXPO_PUBLIC_PORT}/register`, {
       method: "POST",
       headers: {
         "Content-Type": "application/json"
       },
       body: JSON.stringify(payload)
     }).then(res => {
-      console.log("payload received");
       if (!res.ok) { // Login failed
         res.text().then(text => {
           Alert.alert('', text, [{
