--- mobile/app/(tabs)/matches.js
@@ -1,5 +1,0 @@
-export default function Matches() {
-  return (
-    null
-  );
-}
