+++ mobile/app/organizations/[org]/(tabs)/settings.js
@@ -1,0 +1,5 @@
+export default function Settings() {
+  return (
+    null
+  );
+}
